refactor: fix typo
