+++ packages/models/src/utils/index.ts
@@ -1,4 +1,4 @@
 export * from './constant'
 export * from './decorator'
 export * from './exception'
-export * from './heler'
+export * from './helper'
